[Feat] 기록 수정 화면 - 기록 데이터 UI에 반영

+++ Mapmo/Util/Extensions/Double+Extension.swift
@@ -1,5 +1,5 @@
 //
-//  Int+Extension.swift
+//  Double+Extension.swift
 //  Mapmo
 //
 //  Created by 조유진 on 3/18/24.
@@ -7,8 +7,8 @@
 
 import Foundation
 
-extension Int {
+extension Double {
     func convertToCoordinate() -> Double {
-        return Double(self) / 10_000_000.0
+        return self / 10_000_000.0
     }
 }
